requirements.txt
numpy
matplotlib
imageio
imageio-ffmpeg
pyyaml
tqdm
